+++ resources/views/eunomia/modulos/listado_modulos.blade.php
@@ -5,7 +5,9 @@
         Listado
         <small>Módulos</small>
     </h1>
-    <h2>{{ link_to_route('modulos.create', 'Nuevo', null, array('class' => 'btn btn-block btn-success btn-xs')) }}</h2>
+    @if( \Auth::user()->compruebaSeguridad('crear-modulo') == true)
+        <h2>{{ link_to_route('modulos.create', 'Nuevo', null, array('class' => 'btn btn-block btn-success btn-xs')) }}</h2>
+    @endif
 
     <ol class="breadcrumb">
         <li><a href="/admin"><i class="fa fa-dashboard"></i> Home</a></li>
@@ -36,19 +38,19 @@
 
                             <tr>
                                 <td>{{$modulo->nombre}}</td>
-                                </td>
-                                <td>{{ link_to_route('modulos.edit', 'Editar', $modulo, array('class' => 'btn btn btn-warning btn-xs')) }}
-                                    {{ Form::open(array('method'=> 'DELETE', 'route' => array('modulos.destroy', $modulo->id),'style'=>'display:inline','class'=>'form_eliminar')) }}
-                                    {{ Form::submit('Eliminar', array('class' => 'btn btn btn-danger btn-xs')) }}
-                                    {{ Form::close() }}
+                                <td>@if( \Auth::user()->compruebaSeguridad('editar-modulo') == true)
+                                        {{ link_to_route('modulos.edit', 'Editar', $modulo, array('class' => 'btn btn btn-warning btn-xs')) }}
+                                    @endif
+                                    @if( \Auth::user()->compruebaSeguridad('eliminar-modulo') == true)
+                                        {{ Form::open(array('method'=> 'DELETE', 'route' => array('modulos.destroy', $modulo->id),'style'=>'display:inline','class'=>'form_eliminar')) }}
+                                        {{ Form::submit('Eliminar', array('class' => 'btn btn btn-danger btn-xs')) }}
+                                        {{ Form::close() }}
+                                    @endif
 
                                 </td>
                             </tr>
 
                         @endforeach
-
-
-
 
                         </tbody>
